requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
web3==6.13.0
httpx==0.25.2
pydantic==2.5.2
jinja2==3.1.2
python-dotenv==1.0.0
